+++ src/index.ts
@@ -144,23 +144,7 @@ const createRserve = async (
           if (err) {
             reject(err);
           } else {
-            const ocapFuns = Object.fromEntries(
-              Object.entries(data.data).map(([key, fun]) => [
-                key,
-                (...args: any[]) => {
-                  return new Promise((resolve, reject) => {
-                    fun(...args, (err: string, data: unknown) => {
-                      if (err) {
-                        reject(err);
-                      } else {
-                        resolve(data);
-                      }
-                    });
-                  });
-                },
-              ])
-            );
-            // resolve(ocapFuns as any);
+            const ocapFuns = data.data;
             if (schema) resolve(z.object(schema).parse(ocapFuns));
             else resolve(ocapFuns as any);
           }
@@ -182,3 +166,6 @@ const RserveClient = {
 };
 
 export default RserveClient;
+export * as Robj from "./types";
+export * as Rfmt from "./helpers";
+export type * from "./types";